+++ jest.config.js
@@ -1,12 +1,4 @@
 module.exports = {
-  clearMocks: true,
-  setupFiles: ['./testSetup.ts'],
-  preset: 'ts-jest',
-  testMatch: ['**/__tests__/**/*test.(js|ts)'],
-  testPathIgnorePatterns: ['<rootDir>/dist/', '<rootDir>/node_modules/'],
-  globals: {
-    'ts-jest': {
-      isolatedModules: true,
-    },
-  },
+  ...require('@jupiterone/integration-sdk-dev-tools/config/jest'),
+  setupFiles: ['dotenv/config', './jest.env.js'],
 };