+++ src/converter/relationships.ts
@@ -1,23 +1,24 @@
 import {
-  createIntegrationRelationship,
-  Entity,
-  Relationship,
-  RelationshipDirection,
   convertProperties,
+  createMappedRelationship,
+  Entity,
+  MappedRelationship,
+  RelationshipClass,
+  RelationshipDirection,
 } from '@jupiterone/integration-sdk-core';
 
 export const DEVICE_MANAGEMENT_RELATIONSHIP =
   'snipeit_account_manages_hardware';
 export const DEVICE_LOCATION_RELATIONSHIP = 'site_has_hardware';
 
-export const mapHardwareRelationship = (
+export function mapHardwareRelationship(
   account: Entity,
   hardware: Entity,
   filterKey: string,
-): Relationship =>
-  createIntegrationRelationship({
+): MappedRelationship {
+  return createMappedRelationship({
     _key: `${account._key}|manages|${hardware._key}`,
-    _class: 'MANAGES',
+    _class: RelationshipClass.MANAGES,
     _type: DEVICE_MANAGEMENT_RELATIONSHIP,
     _mapping: {
       relationshipDirection: RelationshipDirection.FORWARD,
@@ -31,13 +32,14 @@ export const mapHardwareRelationship = (
       },
     },
   });
+}
 
-export const mapHardwareLocationRelationship = (
+export function mapHardwareLocationRelationship(
   hardware: Entity,
-): Relationship =>
-  createIntegrationRelationship({
+): MappedRelationship {
+  return createMappedRelationship({
     _key: `location:${hardware.locationId}|has|${hardware._key}`,
-    _class: 'HAS',
+    _class: RelationshipClass.HAS,
     _type: DEVICE_LOCATION_RELATIONSHIP,
     _mapping: {
       relationshipDirection: RelationshipDirection.FORWARD,
@@ -46,8 +48,9 @@ export const mapHardwareLocationRelationship = (
       targetEntity: {
         _class: 'Device',
         id: hardware.id,
-        locationId: hardware.locationId,
+        locationId: hardware.locationId as number,
       },
       skipTargetCreation: true,
     },
   });
+}