+++ src/steps/index.ts
@@ -1,0 +1,6 @@
+import { accountSteps } from './account';
+import { hardwareSteps } from './fetch-hardware';
+
+const integrationSteps = [...accountSteps, ...hardwareSteps];
+
+export { integrationSteps };