+++ src/types.ts
@@ -1,4 +1,4 @@
-export interface IntegrationConfig {
+export type IntegrationConfig = {
   hostname: string;
   apiToken: string;
-}
+};