+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+- Upgrade to latest managed integration patterns and libraries
+
 ## 0.1.0 - 2020-05-18
 
 ### Added